requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-ollama
# unstructured
# onnxruntime
chromadb
pymupdf
pytest
pyyaml
psutil
python-dotenv
requests
tqdm
lxml
sentence-transformers
# hf_xet
huggingface_hub[hf_xet]
numpy
pandas
# sklearn
# matplotlib
# seaborn
# torch
# transformers
beautifulsoup4
selenium

# `pip install -U langchain-chroma`
# `pip install "unstructured[md]"`